Transferindo funções de arquivo
O código ainda não roda, estou no meio do processo.
