+++ additions.py
@@ -7,57 +7,63 @@
 import pygame
 from constants import *
 
-pygame.init()
-
 
 # draw main game board
 def draw_board():
-    for i in range(32):
-        column = i % 4
-        row = i // 4
-        if row % 2 == 0:
-            pygame.draw.rect(screen, 'light gray', [600 - (column * 200), row * 100, 100, 100])
-        else:
-            pygame.draw.rect(screen, 'light gray', [700 - (column * 200), row * 100, 100, 100])
-        pygame.draw.rect(screen, 'gray', [0, 800, WIDTH, 100])
-        pygame.draw.rect(screen, 'gold', [0, 800, WIDTH, 100], 5)
-        pygame.draw.rect(screen, 'gold', [800, 0, 200, HEIGHT], 5)
-        status_text = ['White: Select a Piece to Move!', 'White: Select a Destination!',
-                       'Black: Select a Piece to Move!', 'Black: Select a Destination!']
-        screen.blit(big_font.render(status_text[turn_step], True, 'black'), (20, 820))
-        for i in range(9):
-            pygame.draw.line(screen, 'black', (0, 100 * i), (800, 100 * i), 2)
-            pygame.draw.line(screen, 'black', (100 * i, 0), (100 * i, 800), 2)
-        screen.blit(medium_font.render('FORFEIT', True, 'black'), (810, 830))
-        if white_promote or black_promote:
-            pygame.draw.rect(screen, 'gray', [0, 800, WIDTH - 200, 100])
-            pygame.draw.rect(screen, 'gold', [0, 800, WIDTH - 200, 100], 5)
-            screen.blit(big_font.render('Select Piece to Promote Pawn', True, 'black'), (20, 820))
+    colors = ['white', 'light gray']
+    for row in range(8):
+        for col in range(8):
+            # Determine the color of the square
+            color_index = (row + col) % 2
+            color = colors[color_index]
+            # Calculate the position and size of the square
+            square_rect = pygame.Rect(
+                board_x + col * square_size,
+                board_y + row * square_size,
+                square_size,
+                square_size
+            )
+            # Draw the square
+            pygame.draw.rect(screen, color, square_rect)
+    # Draw board outline
+    pygame.draw.rect(screen, 'black', (board_x, board_y, board_width, board_height), 2)
 
 
 # draw pieces onto board
 def draw_pieces():
+    # Draw white pieces
     for i in range(len(white_pieces)):
-        index = piece_list.index(white_pieces[i])
-        if white_pieces[i] == 'pawn':
-            screen.blit(white_pawn, (white_locations[i][0] * 100 + 22, white_locations[i][1] * 100 + 30))
-        else:
-            screen.blit(white_images[index], (white_locations[i][0] * 100 + 10, white_locations[i][1] * 100 + 10))
-        if turn_step < 2:
-            if selection == i:
-                pygame.draw.rect(screen, 'red', [white_locations[i][0] * 100 + 1, white_locations[i][1] * 100 + 1,
-                                                 100, 100], 2)
+        piece = white_pieces[i]
+        position = white_locations[i]
+        image = white_images[piece]
+        col, row = position
+        # Calculate the top-left pixel position of the square
+        pos_x = board_x + col * square_size
+        pos_y = board_y + row * square_size
+        # Center the piece image within the square
+        piece_rect = image.get_rect(center=(
+            pos_x + square_size / 2,
+            pos_y + square_size / 2
+        ))
+        screen.blit(image, piece_rect)
+        if turn_step < 2 and selection == i:
+            pygame.draw.rect(screen, 'red', piece_rect, 2)
 
+    # Draw black pieces
     for i in range(len(black_pieces)):
-        index = piece_list.index(black_pieces[i])
-        if black_pieces[i] == 'pawn':
-            screen.blit(black_pawn, (black_locations[i][0] * 100 + 22, black_locations[i][1] * 100 + 30))
-        else:
-            screen.blit(black_images[index], (black_locations[i][0] * 100 + 10, black_locations[i][1] * 100 + 10))
-        if turn_step >= 2:
-            if selection == i:
-                pygame.draw.rect(screen, 'blue', [black_locations[i][0] * 100 + 1, black_locations[i][1] * 100 + 1,
-                                                  100, 100], 2)
+        piece = black_pieces[i]
+        position = black_locations[i]
+        image = black_images[piece]
+        col, row = position
+        pos_x = board_x + col * square_size
+        pos_y = board_y + row * square_size
+        piece_rect = image.get_rect(center=(
+            pos_x + square_size / 2,
+            pos_y + square_size / 2
+        ))
+        screen.blit(image, piece_rect)
+        if turn_step >= 2 and selection == i:
+            pygame.draw.rect(screen, 'blue', piece_rect, 2)
 
 
 # function to check all pieces valid options on board
@@ -255,50 +261,63 @@ def check_valid_moves():
 
 # draw valid moves on screen
 def draw_valid(moves):
-    if turn_step < 2:
-        color = 'red'
-    else:
-        color = 'blue'
-    for i in range(len(moves)):
-        pygame.draw.circle(screen, color, (moves[i][0] * 100 + 50, moves[i][1] * 100 + 50), 5)
-
+    color = 'red' if turn_step < 2 else 'blue'
+    for move in moves:
+        col, row = move
+        center_x = board_x + col * square_size + square_size / 2
+        center_y = board_y + row * square_size + square_size / 2
+        pygame.draw.circle(screen, color, (center_x, center_y), 5)
 
 # draw captured pieces on side of screen
 def draw_captured():
     for i in range(len(captured_pieces_white)):
         captured_piece = captured_pieces_white[i]
         index = piece_list.index(captured_piece)
-        screen.blit(small_black_images[index], (825, 5 + 50 * i))
+        position = (
+            captured_white_start_x,
+            captured_white_start_y + captured_piece_spacing * i
+        )
+        screen.blit(small_black_images[index], position)
     for i in range(len(captured_pieces_black)):
         captured_piece = captured_pieces_black[i]
         index = piece_list.index(captured_piece)
-        screen.blit(small_white_images[index], (925, 5 + 50 * i))
+        position = (
+            captured_black_start_x,
+            captured_black_start_y + captured_piece_spacing * i
+        )
+        screen.blit(small_white_images[index], position)
 
 
 # draw a flashing square around king if in check
 def draw_check():
-    global check
-    check = False
     if turn_step < 2:
         if 'king' in white_pieces:
             king_index = white_pieces.index('king')
             king_location = white_locations[king_index]
             for i in range(len(black_options)):
                 if king_location in black_options[i]:
-                    check = True
                     if counter < 15:
-                        pygame.draw.rect(screen, 'dark red', [white_locations[king_index][0] * 100 + 1,
-                                                              white_locations[king_index][1] * 100 + 1, 100, 100], 5)
+                        col, row = king_location
+                        pygame.draw.rect(screen, 'dark red', [
+                            board_x + col * square_size,
+                            board_y + row * square_size,
+                            square_size,
+                            square_size
+                        ], 5)
     else:
         if 'king' in black_pieces:
             king_index = black_pieces.index('king')
             king_location = black_locations[king_index]
             for i in range(len(white_options)):
                 if king_location in white_options[i]:
-                    check = True
                     if counter < 15:
-                        pygame.draw.rect(screen, 'dark blue', [black_locations[king_index][0] * 100 + 1,
-                                                               black_locations[king_index][1] * 100 + 1, 100, 100], 5)
+                        col, row = king_location
+                        pygame.draw.rect(screen, 'dark blue', [
+                            board_x + col * square_size,
+                            board_y + row * square_size,
+                            square_size,
+                            square_size
+                        ], 5)
 
 
 def draw_game_over():
@@ -449,164 +468,166 @@ def check_promo_select():
 
 
 # main game loop
-black_options = check_options(black_pieces, black_locations, 'black')
-white_options = check_options(white_pieces, white_locations, 'white')
-run = True
-while run:
-    timer.tick(fps)
-    if counter < 30:
-        counter += 1
-    else:
-        counter = 0
-    screen.fill('dark gray')
-    draw_board()
-    draw_pieces()
-    draw_captured()
-    draw_check()
-    if not game_over:
-        white_promote, black_promote, promo_index = check_promotion()
-        if white_promote or black_promote:
-            draw_promotion()
-            check_promo_select()
-    if selection != 100:
-        valid_moves = check_valid_moves()
-        draw_valid(valid_moves)
-        if selected_piece == 'king':
-            draw_castling(castling_moves)
-    # event handling
-    for event in pygame.event.get():
-        if event.type == pygame.QUIT:
-            run = False
-        if event.type == pygame.MOUSEBUTTONDOWN and event.button == 1 and not game_over:
-            x_coord = event.pos[0] // 100
-            y_coord = event.pos[1] // 100
-            click_coords = (x_coord, y_coord)
-            if turn_step <= 1:
-                if click_coords == (8, 8) or click_coords == (9, 8):
-                    winner = 'black'
-                if click_coords in white_locations:
-                    selection = white_locations.index(click_coords)
-                    # check what piece is selected, so you can only draw castling moves if king is selected
-                    selected_piece = white_pieces[selection]
-                    if turn_step == 0:
-                        turn_step = 1
-                if click_coords in valid_moves and selection != 100:
-                    white_ep = check_ep(white_locations[selection], click_coords)
-                    white_locations[selection] = click_coords
-                    white_moved[selection] = True
-                    if click_coords in black_locations:
-                        black_piece = black_locations.index(click_coords)
-                        captured_pieces_white.append(black_pieces[black_piece])
-                        if black_pieces[black_piece] == 'king':
-                            winner = 'white'
-                        black_pieces.pop(black_piece)
-                        black_locations.pop(black_piece)
-                        black_moved.pop(black_piece)
-                    # adding check if en passant pawn was captured
-                    if click_coords == black_ep:
-                        black_piece = black_locations.index((black_ep[0], black_ep[1] - 1))
-                        captured_pieces_white.append(black_pieces[black_piece])
-                        black_pieces.pop(black_piece)
-                        black_locations.pop(black_piece)
-                        black_moved.pop(black_piece)
-                    black_options = check_options(black_pieces, black_locations, 'black')
-                    white_options = check_options(white_pieces, white_locations, 'white')
-                    turn_step = 2
-                    selection = 100
-                    valid_moves = []
-                # add option to castle
-                elif selection != 100 and selected_piece == 'king':
-                    for q in range(len(castling_moves)):
-                        if click_coords == castling_moves[q][0]:
-                            white_locations[selection] = click_coords
-                            white_moved[selection] = True
-                            if click_coords == (1, 0):
-                                rook_coords = (0, 0)
-                            else:
-                                rook_coords = (7, 0)
-                            rook_index = white_locations.index(rook_coords)
-                            white_locations[rook_index] = castling_moves[q][1]
-                            black_options = check_options(black_pieces, black_locations, 'black')
-                            white_options = check_options(white_pieces, white_locations, 'white')
-                            turn_step = 2
-                            selection = 100
-                            valid_moves = []
-            if turn_step > 1:
-                if click_coords == (8, 8) or click_coords == (9, 8):
-                    winner = 'white'
-                if click_coords in black_locations:
-                    selection = black_locations.index(click_coords)
-                    # check what piece is selected, so you can only draw castling moves if king is selected
-                    selected_piece = black_pieces[selection]
-                    if turn_step == 2:
-                        turn_step = 3
-                if click_coords in valid_moves and selection != 100:
-                    black_ep = check_ep(black_locations[selection], click_coords)
-                    black_locations[selection] = click_coords
-                    black_moved[selection] = True
+if __name__ == "__main__":
+    pygame.init()
+    black_options = check_options(black_pieces, black_locations, 'black')
+    white_options = check_options(white_pieces, white_locations, 'white')
+    run = True
+    while run:
+        timer.tick(fps)
+        if counter < 30:
+            counter += 1
+        else:
+            counter = 0
+        screen.fill('dark gray')
+        draw_board()
+        draw_pieces()
+        draw_captured()
+        draw_check()
+        if not game_over:
+            white_promote, black_promote, promo_index = check_promotion()
+            if white_promote or black_promote:
+                draw_promotion()
+                check_promo_select()
+        if selection != 100:
+            valid_moves = check_valid_moves()
+            draw_valid(valid_moves)
+            if selected_piece == 'king':
+                draw_castling(castling_moves)
+        # event handling
+        for event in pygame.event.get():
+            if event.type == pygame.QUIT:
+                run = False
+            if event.type == pygame.MOUSEBUTTONDOWN and event.button == 1 and not game_over:
+                x_coord = event.pos[0] // 100
+                y_coord = event.pos[1] // 100
+                click_coords = (x_coord, y_coord)
+                if turn_step <= 1:
+                    if click_coords == (8, 8) or click_coords == (9, 8):
+                        winner = 'black'
                     if click_coords in white_locations:
-                        white_piece = white_locations.index(click_coords)
-                        captured_pieces_black.append(white_pieces[white_piece])
-                        if white_pieces[white_piece] == 'king':
-                            winner = 'black'
-                        white_pieces.pop(white_piece)
-                        white_locations.pop(white_piece)
-                        white_moved.pop(white_piece)
-                    if click_coords == white_ep:
-                        white_piece = white_locations.index((white_ep[0], white_ep[1] + 1))
-                        captured_pieces_black.append(white_pieces[white_piece])
-                        white_pieces.pop(white_piece)
-                        white_locations.pop(white_piece)
-                        white_moved.pop(white_piece)
-                    black_options = check_options(black_pieces, black_locations, 'black')
-                    white_options = check_options(white_pieces, white_locations, 'white')
+                        selection = white_locations.index(click_coords)
+                        # check what piece is selected, so you can only draw castling moves if king is selected
+                        selected_piece = white_pieces[selection]
+                        if turn_step == 0:
+                            turn_step = 1
+                    if click_coords in valid_moves and selection != 100:
+                        white_ep = check_ep(white_locations[selection], click_coords)
+                        white_locations[selection] = click_coords
+                        white_moved[selection] = True
+                        if click_coords in black_locations:
+                            black_piece = black_locations.index(click_coords)
+                            captured_pieces_white.append(black_pieces[black_piece])
+                            if black_pieces[black_piece] == 'king':
+                                winner = 'white'
+                            black_pieces.pop(black_piece)
+                            black_locations.pop(black_piece)
+                            black_moved.pop(black_piece)
+                        # adding check if en passant pawn was captured
+                        if click_coords == black_ep:
+                            black_piece = black_locations.index((black_ep[0], black_ep[1] - 1))
+                            captured_pieces_white.append(black_pieces[black_piece])
+                            black_pieces.pop(black_piece)
+                            black_locations.pop(black_piece)
+                            black_moved.pop(black_piece)
+                        black_options = check_options(black_pieces, black_locations, 'black')
+                        white_options = check_options(white_pieces, white_locations, 'white')
+                        turn_step = 2
+                        selection = 100
+                        valid_moves = []
+                    # add option to castle
+                    elif selection != 100 and selected_piece == 'king':
+                        for q in range(len(castling_moves)):
+                            if click_coords == castling_moves[q][0]:
+                                white_locations[selection] = click_coords
+                                white_moved[selection] = True
+                                if click_coords == (1, 0):
+                                    rook_coords = (0, 0)
+                                else:
+                                    rook_coords = (7, 0)
+                                rook_index = white_locations.index(rook_coords)
+                                white_locations[rook_index] = castling_moves[q][1]
+                                black_options = check_options(black_pieces, black_locations, 'black')
+                                white_options = check_options(white_pieces, white_locations, 'white')
+                                turn_step = 2
+                                selection = 100
+                                valid_moves = []
+                if turn_step > 1:
+                    if click_coords == (8, 8) or click_coords == (9, 8):
+                        winner = 'white'
+                    if click_coords in black_locations:
+                        selection = black_locations.index(click_coords)
+                        # check what piece is selected, so you can only draw castling moves if king is selected
+                        selected_piece = black_pieces[selection]
+                        if turn_step == 2:
+                            turn_step = 3
+                    if click_coords in valid_moves and selection != 100:
+                        black_ep = check_ep(black_locations[selection], click_coords)
+                        black_locations[selection] = click_coords
+                        black_moved[selection] = True
+                        if click_coords in white_locations:
+                            white_piece = white_locations.index(click_coords)
+                            captured_pieces_black.append(white_pieces[white_piece])
+                            if white_pieces[white_piece] == 'king':
+                                winner = 'black'
+                            white_pieces.pop(white_piece)
+                            white_locations.pop(white_piece)
+                            white_moved.pop(white_piece)
+                        if click_coords == white_ep:
+                            white_piece = white_locations.index((white_ep[0], white_ep[1] + 1))
+                            captured_pieces_black.append(white_pieces[white_piece])
+                            white_pieces.pop(white_piece)
+                            white_locations.pop(white_piece)
+                            white_moved.pop(white_piece)
+                        black_options = check_options(black_pieces, black_locations, 'black')
+                        white_options = check_options(white_pieces, white_locations, 'white')
+                        turn_step = 0
+                        selection = 100
+                        valid_moves = []
+                    # add option to castle
+                    elif selection != 100 and selected_piece == 'king':
+                        for q in range(len(castling_moves)):
+                            if click_coords == castling_moves[q][0]:
+                                black_locations[selection] = click_coords
+                                black_moved[selection] = True
+                                if click_coords == (1, 7):
+                                    rook_coords = (0, 7)
+                                else:
+                                    rook_coords = (7, 7)
+                                rook_index = black_locations.index(rook_coords)
+                                black_locations[rook_index] = castling_moves[q][1]
+                                black_options = check_options(black_pieces, black_locations, 'black')
+                                white_options = check_options(white_pieces, white_locations, 'white')
+                                turn_step = 0
+                                selection = 100
+                                valid_moves = []
+            if event.type == pygame.KEYDOWN and game_over:
+                if event.key == pygame.K_RETURN:
+                    game_over = False
+                    winner = ''
+                    white_pieces = ['rook', 'knight', 'bishop', 'king', 'queen', 'bishop', 'knight', 'rook',
+                                    'pawn', 'pawn', 'pawn', 'pawn', 'pawn', 'pawn', 'pawn', 'pawn']
+                    white_locations = [(0, 0), (1, 0), (2, 0), (3, 0), (4, 0), (5, 0), (6, 0), (7, 0),
+                                    (0, 1), (1, 1), (2, 1), (3, 1), (4, 1), (5, 1), (6, 1), (7, 1)]
+                    white_moved = [False, False, False, False, False, False, False, False,
+                                False, False, False, False, False, False, False, False]
+                    black_pieces = ['rook', 'knight', 'bishop', 'king', 'queen', 'bishop', 'knight', 'rook',
+                                    'pawn', 'pawn', 'pawn', 'pawn', 'pawn', 'pawn', 'pawn', 'pawn']
+                    black_locations = [(0, 7), (1, 7), (2, 7), (3, 7), (4, 7), (5, 7), (6, 7), (7, 7),
+                                    (0, 6), (1, 6), (2, 6), (3, 6), (4, 6), (5, 6), (6, 6), (7, 6)]
+                    black_moved = [False, False, False, False, False, False, False, False,
+                                False, False, False, False, False, False, False, False]
+                    captured_pieces_white = []
+                    captured_pieces_black = []
                     turn_step = 0
                     selection = 100
                     valid_moves = []
-                # add option to castle
-                elif selection != 100 and selected_piece == 'king':
-                    for q in range(len(castling_moves)):
-                        if click_coords == castling_moves[q][0]:
-                            black_locations[selection] = click_coords
-                            black_moved[selection] = True
-                            if click_coords == (1, 7):
-                                rook_coords = (0, 7)
-                            else:
-                                rook_coords = (7, 7)
-                            rook_index = black_locations.index(rook_coords)
-                            black_locations[rook_index] = castling_moves[q][1]
-                            black_options = check_options(black_pieces, black_locations, 'black')
-                            white_options = check_options(white_pieces, white_locations, 'white')
-                            turn_step = 0
-                            selection = 100
-                            valid_moves = []
-        if event.type == pygame.KEYDOWN and game_over:
-            if event.key == pygame.K_RETURN:
-                game_over = False
-                winner = ''
-                white_pieces = ['rook', 'knight', 'bishop', 'king', 'queen', 'bishop', 'knight', 'rook',
-                                'pawn', 'pawn', 'pawn', 'pawn', 'pawn', 'pawn', 'pawn', 'pawn']
-                white_locations = [(0, 0), (1, 0), (2, 0), (3, 0), (4, 0), (5, 0), (6, 0), (7, 0),
-                                   (0, 1), (1, 1), (2, 1), (3, 1), (4, 1), (5, 1), (6, 1), (7, 1)]
-                white_moved = [False, False, False, False, False, False, False, False,
-                               False, False, False, False, False, False, False, False]
-                black_pieces = ['rook', 'knight', 'bishop', 'king', 'queen', 'bishop', 'knight', 'rook',
-                                'pawn', 'pawn', 'pawn', 'pawn', 'pawn', 'pawn', 'pawn', 'pawn']
-                black_locations = [(0, 7), (1, 7), (2, 7), (3, 7), (4, 7), (5, 7), (6, 7), (7, 7),
-                                   (0, 6), (1, 6), (2, 6), (3, 6), (4, 6), (5, 6), (6, 6), (7, 6)]
-                black_moved = [False, False, False, False, False, False, False, False,
-                               False, False, False, False, False, False, False, False]
-                captured_pieces_white = []
-                captured_pieces_black = []
-                turn_step = 0
-                selection = 100
-                valid_moves = []
-                black_options = check_options(black_pieces, black_locations, 'black')
-                white_options = check_options(white_pieces, white_locations, 'white')
+                    black_options = check_options(black_pieces, black_locations, 'black')
+                    white_options = check_options(white_pieces, white_locations, 'white')
 
-    if winner != '':
-        game_over = True
-        draw_game_over()
+        if winner != '':
+            game_over = True
+            draw_game_over()
 
-    pygame.display.flip()
-pygame.quit()
+        pygame.display.flip()
+    pygame.quit()
